+++ wavespectra/core/__init__.py
@@ -1,1 +1,4 @@
 """General core functions and classes."""
+
+# Support legacy module
+from wavespectra.core import utils as misc
